converts to tsx and cleans up
